update skills section for better readability
