+++ nextjs/src/components/About/Skills.tsx
@@ -1,0 +1,22 @@
+"use client";
+import { motion } from "framer-motion";
+
+import { frontendSkills, backendSkills, devOpsSkills, otherSkills } from "@/misc";
+import SkillsCard from "./SkillsCard";
+
+export const Skills = () => {
+    return (
+        <div>
+            <div className="mt-10 sm:mt-20">
+                <h2 className="text-2xl sm:text-4xl tracking-wide font-bold m-auto mb-5">Skills</h2>
+                <motion.div initial="hidden" whileInView="visible" viewport={{ once: true }} />
+
+                {/* Cards collapse on mobile */}
+                <SkillsCard title="Frontend" skills={frontendSkills} />
+                <SkillsCard title="Backend" skills={backendSkills} />
+                <SkillsCard title="DevOps" skills={devOpsSkills} />
+                <SkillsCard title="Other" skills={otherSkills} />
+            </div>
+        </div>
+    );
+};
