+++ nextjs/src/misc.ts
@@ -1,0 +1,79 @@
+export const frontendSkills = [
+    "React",
+    "TypeScript",
+    "JavaScript",
+    "TailwindCSS",
+    "React Query",
+    "Redux (RTK)",
+    "Firebase Auth",
+    "Next.js",
+    "SEO",
+    "Framer Motion",
+    "UI/UX Design",
+    "Playwright",
+    "Cypress",
+    "Jest",
+    "Vitest",
+    "HTML",
+    "CSS"
+];
+
+export const backendSkills = [
+    "TypeScript",
+    "Node.js",
+    "Express",
+    "Fastify",
+    "Prisma",
+    "PostgreSQL",
+    "SQLServer",
+    "MySQL",
+    "SQLite",
+    "MongoDB",
+    "REST API",
+    "GraphQL",
+    "JWT",
+    "Firebase Auth",
+    "Hoppscotch",
+    "Postman"
+];
+
+export const devOpsSkills = [
+    "Docker",
+    "Linux",
+    "CI/CD Pipelines",
+    "GitHub",
+    "GitLab",
+    // "Kubernetes",
+    "NGINX",
+    "Bash",
+    "PowerShell",
+    "Python",
+    "Security",
+    "Server Management",
+    "Domain Management",
+    "SSL Certificates",
+    "AWS",
+    "Google Cloud",
+    "Firebase",
+    "Oracle Cloud"
+];
+
+export const otherSkills = [
+    "Agile",
+    "Scrum",
+    "C++",
+    "C#",
+    "Rust",
+    "Java",
+    "Jetpack Compose",
+    "React Native",
+    "Flutter",
+    "Dart",
+    "Figma",
+    "Team Management",
+    "Leadership",
+    "Communication",
+    "Building PCs",
+    "Networking",
+    "Troubleshooting"
+];
